implementacion de autenticacion con passport-jwt. Implementacion de cart asociado al user loqueado

+++ src/public/js/cart.js
@@ -4,8 +4,42 @@ document.addEventListener('DOMContentLoaded', () => {
     deleteFromCartButtons.forEach(button => {
         button.addEventListener('click', async (event) => {
             const productId = event.target.dataset.pid;
-            const cartId = '66affd4bc723a31ad3519e85'; 
+
             try {
+                // Obtener cartId del usuario actual
+                const currentUserResponse = await fetch('/api/sessions/current', {
+                    method: 'GET',
+                    headers: {
+                      "Content-Type": "application/json"
+                    },
+                    credentials: 'include'  
+                });
+
+                if (!currentUserResponse.ok) {
+                    throw new Error('Error al obtener la sesión del usuario.');
+                }
+
+                const { cartId } = await currentUserResponse.json();
+
+                if (!cartId) {
+                    throw new Error('No se pudo obtener el cartId');
+                }
+
+                // Confirmación antes de eliminar el producto
+                const confirmDelete = await Swal.fire({
+                    title: '¿Estás seguro?',
+                    text: '¿Quieres eliminar este producto del carrito?',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonText: 'Sí, eliminar!',
+                    cancelButtonText: 'Cancelar'
+                });
+
+                if (!confirmDelete.isConfirmed) {
+                    return; 
+                }
+
+                //  solicitud para eliminar el producto del carrito
                 const response = await fetch(`/api/carts/${cartId}/products/${productId}`, {
                     method: 'DELETE',
                     headers: {
@@ -13,9 +47,12 @@ document.addEventListener('DOMContentLoaded', () => {
                     }
                 });
 
-                const result = await response.json();
-                console.log(result)
+                if (!response.ok) {
+                    throw new Error('Error al eliminar el producto del carrito.');
+                }
 
+                const result = await response.json();
+                
                 if (result.result === 'success') {
                     event.target.closest('li').remove();
 
@@ -26,7 +63,7 @@ document.addEventListener('DOMContentLoaded', () => {
                         confirmButtonText: 'OK'
                     });
                 } else {
-                    throw new Error(result.message);
+                    throw new Error(result.message || 'No se pudo eliminar el producto.');
                 }
             } catch (error) {
                 Swal.fire({
